console: Refactor connect in full view error
